+++ src/index.ts
@@ -1,2 +1,2 @@
-export { default as filemakerWebViewerPlugin } from './lib/plugin';
+export { default as fileMakerWebViewerPlugin } from './lib/plugin';
 export type { default as PluginOptions } from './types/options';