+++ CHANGELOG.md
@@ -1,6 +1,20 @@
 # Changelog
 
 
+## [1.0.1] - 2024-09-03
+
+This release includes important quality of life improvements and bug fixes.
+
+### Changed
+
+- __Important__: The main plugin function has been renamed to `fileMakerWebViewerPlugin` for proper casing.
+- The trigger scripts UI has been updated to be more user-friendly.
+
+### Added
+
+- The ability to toggle the trigger scripts UI.
+- A version number in the trigger scripts UI.
+
 ## [1.0.0] - 2024-09-03
 
 Introducing the first release of `vite-plugin-fmwv`! 🎉